tools/distro-scraper/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "distro-scraper"
version = "0.1.0"
description = "Cloud distribution image scraper for Multipass"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiohttp>=3.9.0",
    "python-dateutil>=2.8.0",
    "pydantic>=2.0.0",
]

[project.scripts]
distro-scraper = "scraper.cli:main"

[project.entry-points."dist_scraper.scrapers"]
debian = "scraper.scrapers.debian:DebianScraper"
fedora = "scraper.scrapers.fedora:FedoraScraper"

[tool.setuptools.packages.find]
where = ["."]
include = ["scraper*"]
